# intc: miwu: npcx: Add wui info of registerd callbaock in miwu driver