This is a further discussion from [PR 57914](https://github.com/zephyrproject-rtos/zephyr/pull/57914)

To remove the deviation of GPIO callback function usage, this PR won't use unused bit fields of `pin_mask` to distinguish when a GPIO event occurs in MIWU's ISR. It reserved a look-up table which size is  196 (3x8x8) bytes to store those configurations instead. Considering the single-list for each MIWU group, the following table is the comparison of RAM usage with the version before [PR 57914](https://github.com/zephyrproject-rtos/zephyr/pull/57914) :

|  Version   |   RO |     RW  |  Total (Uint: bytes)|
| ------------- | ------------- |------------- |------------- |
|PR 57914 |   -100 |   +160   | +60
|This PR  |   +60  |  +380   | +440

And here is the comparison of GPIO interrupt latency with the order of the target callback function in the list:
|  Version      |1st item | 8th item |16th item | 24th item|
| ------------- | ------------- |------------- |------------- |------------- |
| Old version  | 15us   |  25us   |   35us    |     45us |
| PR 57914     | 10us    |  20us |  X  | X  | 
|This PR     | 15us   |  25us   | X  |  X  | 

Ps. We have a mistake in calculating the latency of the 8th item of the list for PR 57914. it should be 20us not 12 us. Sorry for that.

On myst Chromebook, there are 24 gpio interrupt sources and the maximum interrupt latency is ~45us in the old miwu driver (before PR 57914). Besides improving gpio interrupt latency, PR 57914 also improves the interrupt latency of generic devices such as tachometer, eSPI VW events, watchdog, and so on. The maximum latency is 20us no matter source is GPIO or generic devices. But the other approaches need to wait for completing the search of gpio callback list. Then, start to search the items of the generic device callback list. 

Hence, an interrupt latency of generic devices needs to add the maximum interrupt latency of gpio in this PR. If users can choose `pin_mask &= ~BIT(x)` instead of `pin_mask = 0` to cancel gpio callback function, we suggest keeping using PR 57914 since its ram usage is acceptable and improves gpio and device interrupt latency both.

Ps. The CPU clock is 15MHz during calculating the interrupt latency.